refactor(sdk)!: make root exports explicit
Move SDK implementation behind explicit root index manifests so package roots are the public API gate.

BREAKING CHANGE: Provider crypto/parsing helpers and core request/header helpers are no longer exported from package roots.
